01-rag-assistant/requirements.txt
# Core dependencies
openai>=1.3.0
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.10

# Vector database
qdrant-client>=1.7.0

# Embeddings
sentence-transformers>=2.2.2

# Web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Frontend
streamlit>=1.28.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.4.0
requests>=2.31.0

# Development tools
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0

# PDF processing
PyPDF2>=3.0.0
python-multipart>=0.0.6
pypdf>=3.15.0
